Reorganise integration tests for apply
This moves the apply tests into an agree directory structure, i.e. apply -> section -> task -> page test file. The tests are also updated in line with the changes to the section/task hierarchy in the previous commit.
