+++ integration_tests/fixtures/applicationData.json
@@ -1,10 +1,10 @@
 {
-  "area-and-funding": {
+  "funding-information": {
     "funding-information": {
       "fundingSource": "benefits"
     }
   },
-  "about-the-person": {
+  "equality-and-diversity-monitoring": {
     "will-answer-equality-questions": {
       "willAnswer": "yes"
     }
