+++ integration_tests/tests/apply/area_and_funding/funding_information/complete_area_and_funding_section.cy.ts
@@ -1,5 +1,5 @@
-//  Feature: Referrer completes 'About the person' section
-//    So that I can complete the 'About the person' section
+//  Feature: Referrer completes area and funding section
+//    So that I can complete the area and funding section
 //    As a referrer
 //    I want to answer questions within that section of the task list
 //
@@ -9,9 +9,8 @@
 //    And I am logged in
 //    And I am viewing the application task list
 //
-//  Scenario: view task listed within the section
-//    Then I see the task listed within the section
-//    And I see that the task has not been started
+//  Scenario: view task status
+//    Then I see that the task has not been started
 //
 //  Scenario: follow link to first task page
 //    When I follow the link within the section
@@ -19,27 +18,27 @@
 //    And the task list page has the expected question and answers
 //
 //  Scenario: answer is enforced
-//    Given I'm on the 'Will answer equality questions?' task page
+//    Given I'm on the Funding information task page
 //    When I try to continue without answering the question
 //    Then I see that an answer is required
 //
 //  Scenario: return to task list using the back button
-//    Given I'm on the 'Will answer equality questions?' task page
+//    Given I'm on the Funding information task page
 //    When I use the back button
 //    Then I'm on the task list page
 //
 //  Scenario: task completed successfully
-//    Given I'm on the 'Will answer equality questions?' task page
+//    Given I'm on the Funding information task page
 //    When I select an option and choose to save and continue
 //    Then I'm returned to the task list
 //    And I see that the task is now complete
 
-import Page from '../../pages/page'
-import TaskListPage from '../../pages/apply/taskListPage'
-import WillAnswerEqualityQuestionsPage from '../../pages/apply/willAnswerEqualityQuestionsPage'
-import { personFactory, applicationFactory } from '../../../server/testutils/factories/index'
+import Page from '../../../../pages/page'
+import TaskListPage from '../../../../pages/apply/taskListPage'
+import FundingInformationPage from '../../../../pages/apply/fundingInformationPage'
+import { personFactory, applicationFactory } from '../../../../../server/testutils/factories/index'
 
-context('Visit "About the person" section', () => {
+context('Visit area and funding section', () => {
   const person = personFactory.build({ name: 'Roger Smith' })
 
   beforeEach(function test() {
@@ -48,15 +47,7 @@ context('Visit "About the person" section', () => {
     cy.task('stubAuthUser')
 
     cy.fixture('applicationData.json').then(applicationData => {
-      const application = applicationFactory.build({
-        id: 'abc123',
-        data: {
-          'area-and-funding': {
-            'funding-information': { fundingSource: 'personalSavings' },
-          },
-        },
-        person,
-      })
+      const application = applicationFactory.build({ id: 'abc123', person })
       application.data = applicationData
       cy.wrap(application).as('application')
       cy.wrap(application.data).as('applicationData')
@@ -66,15 +57,7 @@ context('Visit "About the person" section', () => {
   beforeEach(function test() {
     // And an application exists
     // -------------------------
-    const newApplication = applicationFactory.build({
-      id: 'abc123',
-      data: {
-        'area-and-funding': {
-          'funding-information': { fundingSource: 'personalSavings' },
-        },
-      },
-      person,
-    })
+    const newApplication = applicationFactory.build({ id: 'abc123', person })
     cy.task('stubApplicationGet', { application: newApplication })
     cy.task('stubApplicationUpdate', { application: newApplication })
 
@@ -88,67 +71,58 @@ context('Visit "About the person" section', () => {
     Page.verifyOnPage(TaskListPage)
   })
 
-  // Scenario: view task listed within the section
+  // Scenario: view task status
   // ----------------------------------------------
   it('shows the task listed within the section', () => {
-    // I'm on the task list page
-    cy.get('h2').contains('Application incomplete')
-
-    // I see the expected SECTION
-    cy.get('.app-task-list__section').contains('About the person')
-
-    // I see the expected TASK
-    cy.get('.app-task-list__task-name').contains('Complete equality and diversity monitoring')
-
-    // And I should see that the task has not been started
+    // I see that the task has not been started
     const taskListPage = Page.verifyOnPage(TaskListPage)
-    taskListPage.shouldShowTaskStatus('about-the-person', 'Not started')
+    taskListPage.shouldShowTaskStatus('funding-information', 'Not started')
   })
 
   // Scenario: follow link to first task page
   // ----------------------------------------
   it('offers a link to the first page of the task', function test() {
     // When I follow the link within the section
-    cy.get('a').contains('Complete equality and diversity monitoring').click()
+    cy.get('a').contains('Add funding information').click()
 
     // Then I'm on the expected page
-    const page = Page.verifyOnPage(WillAnswerEqualityQuestionsPage, this.application)
+    Page.verifyOnPage(FundingInformationPage, this.application)
 
-    // And I see the expected caption
-    page.hasCaption()
+    // And the task list page has the expected questions and answers
+    // ------------------------------------------------------------
+    // And I see the expected question
+    cy.get('h1').contains('How will Roger Smith pay for their accommodation and service charge')
 
-    // And the page has the expected questions and answers
-    page.hasQuestionsAndAnswers()
-
-    // And I see that the section is optional
-    page.isOptional()
-
-    // And I see an explanation of the rationale for collecting this information
-    page.hasRationale()
+    // And I see the expected answers
+    cy.get('label').contains('Personal money or savings')
+    cy.get('label').contains('Benefits')
+    cy.get('.govuk-hint.govuk-radios__hint').contains(
+      'This includes Housing Benefit and Universal Credit, Disability Living Allowance, and Employment and Support Allowance',
+    )
   })
 
   // Scenario: answer is enforced
   // ----------------------------
   it('enforces answer', function test() {
-    // Given I'm on the 'Will answer equality questions' task page
-    cy.get('a').contains('Complete equality and diversity monitoring').click()
+    // Given I'm on the Funding information task page
+    cy.get('a').contains('Add funding information').click()
 
     // I try to continue without making a choice
     cy.get('button').contains('Save and continue').click()
 
     // Then I see that an answer is required
-    const page = new WillAnswerEqualityQuestionsPage(this.application)
-    page.shouldShowErrorMessagesForFields(['willAnswer'])
+    const fundingInfoPage = new FundingInformationPage(this.application)
+    fundingInfoPage.shouldShowErrorMessagesForFields(['fundingSource'])
   })
 
   // Scenario: return to task list using the back button
   //----------------------------------------------------
   it('takes me back to the task list page', function test() {
-    // Given I'm on the 'Equality and diversity' task page
-    cy.get('a').contains('Complete equality and diversity monitoring').click()
+    // Given I'm on the Funding information task page
+    cy.get('a').contains('Add funding information').click()
 
     // When I use the back button
-    const page = Page.verifyOnPage(WillAnswerEqualityQuestionsPage, this.application)
+    const page = Page.verifyOnPage(FundingInformationPage, this.application)
     page.clickBack()
 
     // Then I'm on the task list page
@@ -158,12 +132,12 @@ context('Visit "About the person" section', () => {
   // Scenario: task completed successfully
   // -------------------------------------
   it('submits the valid form', function test() {
-    // Given I'm on the 'Equality and diversity' task page
-    cy.get('a').contains('Complete equality and diversity monitoring').click()
-    const page = Page.verifyOnPage(WillAnswerEqualityQuestionsPage, this.application)
+    // Given I'm on the Funding information task page
+    cy.get('a').contains('Add funding information').click()
+    const page = Page.verifyOnPage(FundingInformationPage, this.application)
 
     // When I select an option and click save and continue
-    page.checkRadioButtonFromPageBody('willAnswer')
+    page.checkRadioButtonFromPageBody('fundingSource')
 
     // after submission of the valid form the API will return the answered question
     // -- note that the presence of the _page_ only is required in application.data
@@ -171,11 +145,8 @@ context('Visit "About the person" section', () => {
     const answered = {
       ...this.application,
       data: {
-        'area-and-funding': {
+        'funding-information': {
           'funding-information': {},
-        },
-        'about-the-person': {
-          'will-answer-equality-questions': {},
         },
       },
     }
@@ -187,6 +158,6 @@ context('Visit "About the person" section', () => {
     const taskListPage = Page.verifyOnPage(TaskListPage)
 
     // And I see that the task is now complete
-    taskListPage.shouldShowTaskStatus('about-the-person', 'Completed')
+    taskListPage.shouldShowTaskStatus('funding-information', 'Completed')
   })
 })
